resolving issue

+++ server/src/controllers/user.controller.js
@@ -86,10 +86,9 @@ export const postLogout = async (req, res) => {
 // **User Profile**
 export const getProfile = async (req, res) => {
   try {
-    const id = req.user._id.toString();
-    console.log(req.user._id.toString())
+    const id = req.user._id;
 
-    const user = await User.findById(id).populate("healthPortfolio").select("-password");
+    const user = await HealthPortfolio.findById({userId: id}).populate("userId").select("-password");
     if (!user) return res.status(404).json({ message: "User not found" });
 
     res.status(200).json(user);
@@ -101,7 +100,7 @@ export const getProfile = async (req, res) => {
 export const updateUser = async (req, res) => {
   try {
     const user = req.user;
-    const updatedUser = await User.findByIdAndUpdate(user._id, req.body, { new: true });
+    const updatedUser = await HealthPortfolio.findByIdAndUpdate({userId: user._id}, req.body, { new: true });
     res.status(200).json({ message: "User updated successfully", updatedUser });
   } catch (error) {
     res.status(500).json({ message: "Error updating user", error });
@@ -110,7 +109,7 @@ export const updateUser = async (req, res) => {
 
 export const deleteUser = async (req, res) => {
   try {
-    await User.findByIdAndDelete(req.user._id);
+    await User.findByIdAndDelete({userId: req.user._id});
     res.status(200).json({ message: "User deleted successfully" });
   } catch (error) {
     res.status(500).json({ message: "Error deleting user", error });
